Abstracted presistence tests
