+++ persistence-cassandra/src/main/scala/com/lightbend/lagom/javadsl/persistence/cassandra/testkit/TestUtil.scala
@@ -1,0 +1,15 @@
+/*
+ * Copyright (C) 2016 Lightbend Inc. <http://www.lightbend.com>
+ */
+package com.lightbend.lagom.javadsl.persistence.cassandra.testkit
+
+import com.lightbend.lagom.javadsl.persistence.testkit.AbstractTestUtil
+import com.typesafe.config.Config
+
+object TestUtil extends AbstractTestUtil {
+
+  def persistenceConfig(testName: String, cassandraPort: Int, useServiceLocator: Boolean): Config = {
+    com.lightbend.lagom.javadsl.persistence.testkit.TestUtil.persistenceConfig(testName, cassandraPort, useServiceLocator)
+  }
+
+}
