+++ persistence-cassandra/src/test/scala/com/lightbend/lagom/javadsl/persistence/cassandra/CassandraPersistentEntityActorSpec.scala
@@ -1,0 +1,8 @@
+/*
+ * Copyright (C) 2016 Lightbend Inc. <http://www.lightbend.com>
+ */
+package com.lightbend.lagom.javadsl.persistence.cassandra
+
+import com.lightbend.lagom.javadsl.persistence.AbstractPersistentEntityActorSpec
+
+class CassandraPersistentEntityActorSpec extends CassandraPersistenceSpec with AbstractPersistentEntityActorSpec
