+++ persistence-cassandra/src/test/scala/com/lightbend/lagom/javadsl/persistence/cassandra/testkit/EmbeddedCassandraPersistentActorSpec.scala
@@ -1,0 +1,9 @@
+/*
+ * Copyright (C) 2016 Lightbend Inc. <http://www.lightbend.com>
+ */
+package com.lightbend.lagom.javadsl.persistence.cassandra.testkit
+
+import com.lightbend.lagom.javadsl.persistence.cassandra.CassandraPersistenceSpec
+import com.lightbend.lagom.javadsl.persistence.testkit.AbstractEmbeddedPersistentActorSpec
+
+class EmbeddedCassandraPersistentActorSpec extends CassandraPersistenceSpec with AbstractEmbeddedPersistentActorSpec
